+++ waypoints/src/main/kotlin/de/md5lukas/waypoints/WaypointsPlugin.kt
@@ -354,8 +354,6 @@ class WaypointsPlugin : JavaPlugin() {
         SimplePie("particle_pointer_enabled") {
           waypointsConfig.pointers.particle.enabled.toString()
         })
-    metrics.addCustomChart(
-        SimplePie("trails_pointer_enabled") { waypointsConfig.pointers.trail.enabled.toString() })
   }
 
   private fun startBackgroundTasks() {